gschem: Refactor resnap logic to ensure correctness.
Previously, there were code paths in the resnap logic which could lead
to accessing uninitialised memory (reported on IRC by Gareth Edwards).
The algorithm that chose whether to resnap or not also iterated over
the selection twice, when once is actually enough.

This patch changes the resnap logic to decide on snappability in one
pass over the selection, ensuring that resnap only occurs if (1) there
is at most one object selected that isn't an attribute, and (2) all
attribute objects are attached to the same object (whether it's
selected or not).