gnetlist: Improved checking of attribute identity.
gnetlist warns the user if several components with the same refdes have
same named attributes with different values. So far, it also yielded the
warning in the case when some of such components missed the attribute in
question. There is no point to have, say, footprint= attached on several
slots of a slotted part, so this check is now omitted for symbol
instances having no requested attribute.

This change is not backward compatible with previous gnetlist behaviour.