Use newer `ustar' format for distribution
By default, the old v7 tar format is used for creating the distribution
tarball.  This format has severe limitations (only regular files,
hardlinks and symlinks; pathname up to 99 characters; file size up to
8 GB; user and group stored numerically).  This causes a gnetlist DRC2
test file with a 101-character pathname to be dropped from the archive
(a warning message is printed about this, but it's easy to miss in the
build scroll), which in turn causes `make distcheck' to fail.

The newer `ustar' format is widely supported today:
https://dev.gentoo.org/~mgorny/articles/portability-of-tar-features.html