gschem: Make attaching attributes work as expected
"Attributes / Attach Attributes" used to depend on the order in which
objects were selected, taking the first selected object as the base
object to attach attributes to, regardless of its type.  Now, the
selected objects are searched for a base object of suitable type, and
only if exactly one is found, the attributes are attached to it.