gnetlist-legacy: Remove test section "renamed-nets"
The concept of "renaming" nets doesn't exist in the current version of
gnetlist any more.  The Python version of the `geda' backend doesn't
output a "renamed-nets" section, but the Guile version attempts to call
"geda:renamed-nets" which fails unless the backend is executed via the
gnetlist-legacy binary.

Since users tend to use this backend for debugging and experimentation
purposes and it's unlikely that the net renaming mechanism will break
in an unmaintained codebase without causing other observable problems,
simply removing the tests appears to be the best option.