Fix licensing inconsistencies
Some files in the source tree state that they are licensed under the
GNU LGPL.  For files included from external libraries (base64, rsvg),
this makes sense; but there are some random files in libgeda, gschem,
and gsymcheck which state they are licensed under the GNU LGPL, as well.

Since the majority of files in gEDA/gaf is licensed under the GNU GPL
and some files (e.g., libgeda/src/s_path.c) explicitly state that they
are based on LGPL'd code and an exception to the rule that gEDA/gaf,
as a whole, is licensed under the GNU GPL, I assume this is an error
and fixed the licensing information for these files.