Migrate to Guile 2.2
In Guile 2.2, the API function `scm_frame_procedure' has been replaced
with `scm_frame_procedure_name', so only the name of the currently
executed procedure can now be retrieved, not the procedure itself.

Instead of using the procedure itself as a look-up key for the address
of the corresponding Python callable, we now have to generate a unique
name for each Guile procedure and use that as the key.