make gschem update the pin #s for bypass sab-param when changing slot
1. Added isolated SAB code. Given a component object the code extracts
all the sab-param attributes and parses them (ignoring any that are
malformed) producing a data structure. The contents of that data structure
can then be modified. After modification the structure can be converted
back into sab-param attributes which replace the source attributes. Any
attributes that were ignored for being malformed are left untouched.
Currently the only supported modification is to replace bypass pin #s.

	modified:   libgeda/include/Makefile.am
	new file:   libgeda/include/libgeda/sab.h
	modified:   libgeda/src/Makefile.am
	new file:   libgeda/src/sab.c

2. Add calls to the SAB system to s_slot_update_object so the bypass
pin numbers are updated when the slot is changed. Note when manually
changing the slot number the bypass pin changes are reflected in the
sab-param shown in the schematic but not the one shown in the
'Attributes' panel. That one is updated if the component is deselected
and then reselected causeing the panel to reload the attributes. This
should certainly be fixed but right now I don't have the time to figure
out where the change needs to be made.
Also identified a memory leak so fixed that as well.

	modified:   libgeda/src/s_slot.c

TODO: Need to figure out how to have the sab-param attributes affect
printout. I didn't do it because I really don't need it but mostly
because I couldn't come up with good solutions to:
1) How to specify which context you are producing hardcopy for.
2) How to draw bypass connections on a device with more than two
   pins and/or where the pins are not directly across from each other.
3) How to handle exec actions in hardcopy.

Maybe somebody smarter/more motivated than me can figure out a solution.
Adding the SAB functionality to gschem was really only a convience item anyway.