netlist: Fix namespaces for multi-source subsheets
Components can have multiple source= attributes and thus multiple
subsheets.  However, since the namespace for package and net names
used to be the sheet instance, splitting packages across and
connecting nets between the individual subsheets didn't work.

In order to fix this, use the *instantiating component* instance as
the namespace for package and net names instead.

Reported-by: Klaus Rudolph <lts-rudolph@gmx.de>