gschem: Don't scroll log window unless realized
gtk_text_view_scroll_to_iter doesn't check if the text view has a
layout; if it hasn't, an assertion in gtk_text_layout_get_iter_location
is violated, and the program aborts.  In order to avoid that, check if
the text view is realized before trying to scroll it.