netlist: Allow duplicated pins to stay unconnected
In gEDA/gaf 1.10.0, I changed the handling of identically-numbered pins
in multiple components of a package.  This caused auto-generated nets
on unconnected pins to conflict with the connected net and with each
other unless each pin would be explicitly connected.

Fix this by ignoring unconnected nets if at least one instance of the
pin is connected, and using only a single unconnected net otherwise.

Reported-by: Richard Rasker <rasker@linetec.nl>